+++ src/components/Logo/LogoTypes.tsx
@@ -1,0 +1,32 @@
+export enum LogoColors {
+  Default = "",
+  UiBlack = "ui.black",
+  UiWhite = "ui.white",
+}
+
+export enum LogoSizes {
+  ExtraExtraSmall = "xxsmall",
+  ExtraSmall = "xsmall",
+  Small = "small",
+  Medium = "medium",
+  Large = "large",
+  Default = "default",
+}
+
+export enum LogoNames {
+  BrooklynPublicLibraryBlack = "logo_bpl_black",
+  LPAColor = "logo_lpa_color",
+  LPAWhite = "logo_lpa_white",
+  NYPLBlack = "logo_nypl_full_black",
+  NYPLWhite = "logo_nypl_full_white",
+  NYPLLionBlack = "logo_nypl_lion_black",
+  NYPLLionWhite = "logo_nypl_lion_white",
+  QueensPublicLibraryColor = "logo_qpl_color",
+  QueensPublicLibraryAltBlack = "logo_qpl_alt_black",
+  SchomburgColor = "logo_schomburg_color",
+  SchomburgCircleColor = "logo_schomburg_circle_color",
+  SimplyEColor = "logo_simplye_color",
+  SimplyEBlack = "logo_simplye_black",
+  SNFLBlack = "logo_snfl_black",
+  SNFLWhite = "logo_snfl_white",
+}